Add files via upload
When there are two natural numbers A and B, A%B is the remainder of A divided by B. For example, the remainder of 7, 14, 27, 38 divided by 3 is 1, 2, 0, 2.

After entering 10 numbers, divide them by 42 to get the remainder. Then write a program that prints out how many different values there are.