Add files via upload
Sejun ruined the final exam. Sejun decided to manipulate the score and take it home. First, Sejun chose the maximum value among his scores. This value is called M. Then he fixed all the scores to score/M*100.

For example, if Sejun's highest point was 70 and his math score was 50, his math score would be 50/70*100, resulting in 71.43 points.

When Sejun's grade is newly calculated according to the above method, write a program to find a new average.