Add files via upload
There are results of OX quizzes such as "OOXXOXXOOO". O is the problem right, X is the problem wrong. If a question is correct, the score for that question is the number of consecutive O's up to that question. For example, the score for question 10 is 3.

The score for "OOXXOXXOOO" is 1+2+0+0+1+0+0+1+2+3 = 10 points.

When given the results of the OX quiz, write a program to get points.