Add files via upload
Problem
90% of new college students think they are above average in their class. You have to tell them the sad truth.

Input
In the first line, the number of test cases C is given.
From the second line, for each test case, the number of students N (1 ≤ N ≤ 1000, N is an integer) is given as the first number, followed by N scores. The score is an integer greater than or equal to 0 and less than or equal to 100.

Output
For each case, the percentage of students who exceed the average is rounded up by one line and printed to the third decimal place.