Add files via upload
The self number was established in 1949 by the Indian mathematician D.R. Kaprekar named it. For a positive integer n, let d(n) be defined as a function that adds each digit of n and n. For example, d(75) = 75+7+5 = 87.

Given a positive integer n, starting with this number, we can create an infinite sequence such as n, d(n), d(d(n)), d(d(d(n))), ... have.

For example, if you start with 33, the next number is 33 + 3 + 3 = 39, the next number is 39 + 3 + 9 = 51, and the next number is 51 + 5 + 1 = 57. In this way, you can create a sequence like this:

33, 39, 51, 57, 69, 84, 96, 111, 114, 120, 123, 129, 141, ...

We call n the constructor of d(n). In the above sequence, 33 is a constructor of 39, 39 is a constructor of 51, and 51 is a constructor of 57. In some cases, there are more than one constructor. For example, 101 has two constructors (91 and 100).

A number without a constructor is called a self number. There are a total of 13 self-numbers less than 100. 1, 3, 5, 7, 9, 20, 31, 42, 53, 64, 75, 86, 97

Write a program that prints self-numbers less than or equal to 10000, one per line.