Add files via upload
Problem
If each digit of a positive integer X forms an order sequence, the number is called a single number. An equal order sequence refers to a sequence in which the difference between two consecutive numbers is constant. Given N, write a program that prints the number of numbers greater than or equal to 1 and less than or equal to N.

Input
In the first line, a natural number N less than or equal to 1,000 is given.

Output
In the first line, the number of numbers greater than or equal to 1 and less than or equal to N is printed.